Linux Vulkan port: full upstream-architecture port renders 3D + 2D
Phase 3 of the BetterVR Linux port. Replaces the D3D12 composer with a
Vulkan one (`vk_composer.{h,cpp}`), ports the rendering pipeline
(swapchain, texture, openxr.cpp, renderer) to Vulkan equivalents, adds
cross-instance image/semaphore sharing via VK_KHR_external_memory_fd /
VK_KHR_external_semaphore_fd, and runtime-compiles GLSL shaders with
shaderc instead of HLSL. Win32-only and D3D12-only code is gated behind
new compile feature flags / a Linux compat header.

Notable bug fixes uncovered during bring-up:
- Drop the HLSL-style Y-flip in the present vertex shader. Vulkan's
  framebuffer is y-down natively, so the negation was a double-flip:
  display rendered upside-down AND triangle winding became CCW under
  Vulkan's y-up area-formula convention, which `frontFace=CW` +
  `cullMode=BACK` then culled, producing a black 3D layer with no
  fragments.
- Track all VkImages in `imageResolutions`, not just the >=1280x720
  ones. The previous size filter dropped BotW's 3D depth image on
  Linux, breaking format checks downstream.
- Re-validate format on every magic-clear instead of latching the
  first matching image to `s_curr3DColorImage` / `s_curr3DDepthImage`.
  BotW uses multiple per-slot color and depth framebuffers, so the
  latched-handle check rejected ~50% of valid captures.
- Snapshot Is3DComplete/Is2DComplete in `EndFrame` before `Reset()`
  wipes the slot, so the trace log faithfully reflects the state we
  evaluated for slot selection.

+++ src/hooking/skeleton.cpp
@@ -155,7 +155,8 @@ public:
         glm::vec3 ortho = glm::normalize(glm::cross(planeNormal, dirNorm));
 
         // arm 1 direction (world)
-        glm::vec3 arm1Dir = glm::normalize(dirNorm * cos(alpha) + ortho * sin(alpha));
+        // glm vec * float is fine; glm vec * double trips overload resolution on Linux.
+        glm::vec3 arm1Dir = glm::normalize(dirNorm * static_cast<float>(cos(alpha)) + ortho * static_cast<float>(sin(alpha)));
 
         // arm 2 direction (world)
         glm::vec3 elbowPos = rootPos + arm1Dir * l1;
